refactor: move sprint components to separate folder, change some sprintgame logic
