+++ my-app/src/App.tsx
@@ -6,7 +6,7 @@ import Landing from './components/start-page/Landing';
 import { Routes, Route } from 'react-router-dom';
 import { APP_ROUTES } from './utils/Constants';
 import Textbook from './components/textbook/Textbook';
-import Sprint from './components/sprint/Sprint';
+import Sprint from './pages/sprint/Sprint';
 import Audiocall from './components/audiocall/Audiocall';
 import Stat from './components/statistics/Stat';
 import SignInForm from './components/REFACTORING/Authorisation/SignInForm';
