+++ my-app/src/components/REFACTORING/Game/TextTimer.tsx
@@ -1,6 +1,6 @@
 import { Typography } from '@mui/material';
 import { Dispatch, useEffect, useState } from 'react';
-import { InitialState, ReducerAction, SprintActionTypes } from '../../types/sprintTypes';
+import { InitialState, ReducerAction, SprintActionTypes } from '../../../types/sprintTypes';
 
 type Props = {
   quizState: InitialState;
