+++ my-app/src/components/REFACTORING/SprintGame/SprintGameContent.tsx
@@ -1,14 +1,14 @@
 import { useCallback, useEffect, useState } from 'react';
-import { updateUserWord } from '../../services/UserWordService';
-import { API_URL, GAME_TYPE, WORD_STATUS } from '../../utils/Constants';
-import GameControls from '../REFACTORING/Game/GameControls';
-import GameQuestion from '../REFACTORING/Game/GameQuestion';
-import GameBtns from '../REFACTORING/Game/GameBtns';
-import { StyledStack } from '../REFACTORING/Game/styles';
-import correctAudioSrc from '../../assets/correct.mp3';
-import incorrectAudioSrc from '../../assets/incorrect.mp3';
-import { useSprintContext } from '../../store/hooks';
-import { SprintActionTypes } from '../../types/sprintTypes';
+import { updateUserWord } from '../../../services/UserWordService';
+import { API_URL, GAME_TYPE, WORD_STATUS } from '../../../utils/Constants';
+import GameControls from './GameControls';
+import GameQuestion from './GameQuestion';
+import GameBtns from './GameBtns';
+import { StyledStack } from '../Game/styles';
+import correctAudioSrc from '../../../assets/correct.mp3';
+import incorrectAudioSrc from '../../../assets/incorrect.mp3';
+import { useSprintContext } from '../../../store/hooks';
+import { SprintActionTypes } from '../../../types/sprintTypes';
 
 const correctAudio = new Audio(correctAudioSrc);
 const incorrectAudio = new Audio(incorrectAudioSrc);
