+++ my-app/src/components/audiocall/AudioGame.tsx
@@ -2,8 +2,8 @@ import { Container } from '@mui/material';
 import { Dispatch, SetStateAction, useContext, useEffect, useState } from 'react';
 import { updateUserWord } from '../../services/UserWordService';
 import { API_URL, GAME_TYPE, WORD_STATUS } from '../../utils/Constants';
-import GameScore from '../game/GameScore';
-import { GameAnswers } from '../sprint/Sprint';
+import GameScore from '../REFACTORING/Game/GameScore';
+import { GameAnswers } from '../../pages/sprint/Sprint';
 import { AudioWords, AudioContext } from './Audiocall';
 import VolumeUpIcon from '@mui/icons-material/VolumeUp';
 import Button from '@mui/material/Button';
