+++ my-app/src/components/audiocall/Game.tsx
@@ -2,8 +2,8 @@ import { Box } from '@mui/material';
 import React, { useContext, useEffect, useState } from 'react';
 import { APP_ROUTES, GAME_TYPE } from '../../utils/Constants';
 import LevelModal from './LevelModal';
-import { QuizContext } from '../sprint/Sprint';
-import SprintGame from '../sprint/SprintGameContent';
+import { QuizContext } from '../../pages/sprint/Sprint';
+import SprintGame from '../REFACTORING/SprintGame/SprintGameContent';
 // import Timer from './Timer';
 import Timer from '../game/Timer';
 import { LoadingIcon } from '../shared/LoadingIcon';
