--- my-app/src/components/game/Game.module.css
@@ -1,29 +1,0 @@
-.timer {
-  display: flex;
-  /* position: absolute;
-  top: 0;
-  right: 0; */
-  width: 50px;
-  height: 50px;
-  justify-content: center;
-  align-items: center;
-  border-radius: 5px;
-}
-
-.game {
-  position: relative;
-  display: flex;
-  width: 100%; 
-  height: auto;
-  justify-content: center;
-}
-
-.gameLoadingIcon {
-  position: absolute;
-  z-index: 1150;
-  width: auto;
-  height: auto;
-  left: 50%;
-  top: 50%;
-  transform: translate(-50%, -50%);
-}
