--- my-app/src/components/sprint/sprint.module.css
@@ -1,3 +1,0 @@
-.sprintContainer {
-    background-color: yellowgreen;
-}
